Created a composable test result class.
The purpose of this class is to make possible to define all the XML
report generation logic inside a single `TestResult` subclass while
allowing it to be used together with another `TestRunner` / `TestResult`
pair.

For instance, it will be possible to use the default `TextTestRunner`
and `TextTestResult` AND genererate a XML report at the same time.