Create a valid XML in case of writing the output to a single file.
If the output is written to a single file we have to ensure the document is
generated as a whole and the single testsuite must be placed below the
element testsuites. Concating the output is not an option because this breaks
the XML by having more than one tag as first element which is against the
standard.