Add time zone to reported timestamp
When used on a timezone-unaware `datetime`, Python's
`datetime.isoformat()` method produces a timestamp without a timezone
designator. While this is technically allowed by ISO8601, most ISO8601
parsers expect a timezone designator of some sort.

This PR adds an `astimezone()` call before creating the timestamp. When
used on a timezone-unaware `datetime` (as the one here is), it will
assume the `datetime` is relative to the system timezone and append that
to the timestamp. This is the correct behavior in this instance.