Add support for setting custom xml attributes
With the newly added 'annotate' decorator, it is now possible to add custom
xml attributes to both test cases and test suites.
This makes it possible to mark tests with IDs, which can then be used to
automatically process the test results and integrate it with third
party test management tools.