setup.py: drop data_files, installs LICENSE to incorrect place
As reported in issue
https://github.com/xmlrunner/unittest-xml-reporting/issues/284, the
data_files statement in the setup() calls installs the LICENSE file in
the wrong place: in /usr/LICENSE, or even /LICENSE depending on the
configuration.

So let's drop this, and let setuptools install the LICENSE file
automatically: since setuptools v56, the license_files attribute is
automatically assigned to a default value, which includes "LICENSE",
so there is in fact nothing to do to the get the LICENSE file
installed at the correct location. See
https://setuptools.pypa.io/en/latest/history.html#v56-0-0.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>